+++ back-end/test-fns/monitoring.test.js
@@ -5,7 +5,7 @@
 */
 import { test, describe } from '@jest/globals'
 
-import { fns } from '@akauppi/firebase-jest-testing'
+import { fns } from '@akauppi/firebase-jest-testing/cloudFunctions'
 
 describe("monitoring functions", () => {
   const fnLogs = fns.httpsCallable('logs_v190720');