+++ back-end/test-rules/projectsC.test.js
@@ -1,11 +1,9 @@
 /*
 * back-end/test-rules/projectsC.test.js
 */
-import { strict as assert } from 'assert'
 import { test, expect, describe, beforeAll } from '@jest/globals'
 
-import { dbAuth } from '@akauppi/firebase-jest-testing/firestoreTestingReadOnly'
-import { FieldValue } from '@akauppi/firebase-jest-testing/firestoreTesting'
+import { dbAuth, FieldValue } from '@akauppi/firebase-jest-testing/firestoreReadOnly'
 
 const anyDate = new Date();   // a non-server date